update the samples
